**NB**: this is a proof-of-concept PR which only modifies `dequantize_block_q4_0()` (this is the quantization method I'm most familiar with). If there is interest, I will modify all quantization kernels.
**NB2**: I only did rudimentary testing: I checked that `./main -b 512 -t 12 -n 40 -f prompts/dan.txt -m models/13B/ggml-model-q4_0.bin --no-mmap -s 12345` generates the same tokens with and without this diff. Again, if there is interest, I will test the changes more thoroughly.

### High-level overview

`dequantize_block_q4_0()` should be memory-bound, but currently it's not. Here's what [NSight Compute](https://developer.nvidia.com/nsight-compute) thinks:

![Compute/memory utilization](https://user-images.githubusercontent.com/42850/235268468-1ed402bd-a873-4aa0-b745-d0ec46fe8ad8.png)
![Roofline](https://user-images.githubusercontent.com/42850/235268471-a7ad8574-17d4-4b65-a4ab-f8b49d1dfb25.png)

The most obvious problem I'm trying to fix in this PR is that using 1 thread per thread block to load the quantized block bytes is very inefficient. Threads within a thread block are arranged in [32-thread warps](https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#simt-architecture), and all threads in a warp always execute the same instruction simultaneously. So, instead of issuing many sequential loads using only the first thread of the warp, we can make all 32 threads issue different loads to dequantize different floats. These loads will then be [coalesced](https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#device-memory-accesses) whenever possible. The 32 threads will also perform arithmetic in parallel, but since we are memory-bound, it doesn't matter that much.

Other tweaks implemented here (guided by NSight Compute):
  * running only warp per thread block leads to low occupancy and warp scheduling stalls, so I increased the number of threads in the thread block (1024 is the [maximum number of threads](https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#features-and-technical-specifications); smaller values work, too)
  * we know that the number of quantized blocks is [divisible by 2](https://github.com/ggerganov/llama.cpp/blob/master/ggml.c#L2628), and handling 2 blocks per warp results in utilizing memory better
  * NSight Compute showed that the hardware failed to coalesce byte-sized loads for some reason, so I switched the loads to 4-byte-loads. We could even use `uint4` to load the whole block at once, but unfortunately the block bytes are not properly aligned to 16 bytes.

Here's the NSight Compute analysis for the new kernel:

![Compute/memory utilization](https://user-images.githubusercontent.com/42850/235270661-56a7267f-e27a-424b-be27-00f27f721dad.png)
![Roofline](https://user-images.githubusercontent.com/42850/235270662-3778f08d-bc85-486e-a6d4-634473def3c0.png)

We still have a large stall waiting for the result of a 4-byte memory load, but I ~~am out of ideas~~ don't think it's worth chasing it further (see the performance measurements below):
![Screenshot 2023-04-29 at 01 56 14](https://user-images.githubusercontent.com/42850/235271352-54f14bf0-ba32-435d-aa5a-a7a62ed65699.png)

### Performance measurements

These are the results of running `./perplexity -m models/13B/ggml-model-q4_0.bin -f wikitext-2-raw/wiki.test.raw` under `NSight Systems` with `--stats` before the PR on an **A100-SXM4-40GB** GPU:
```
 Time (%)  Total Time (ns)  Instances   Avg (ns)     Med (ns)    Min (ns)   Max (ns)   StdDev (ns)                             Name
 --------  ---------------  ---------  -----------  -----------  ---------  ---------  -----------  ----------------------------------------------------------
...
     29.0      489,024,865        449  1,089,142.2    644,828.0    608,412  3,791,913    546,249.9  dequantize_block_q4_0(const void *, float *)
...
```

And after the PR:
```
 Time (%)  Total Time (ns)  Instances   Avg (ns)     Med (ns)    Min (ns)   Max (ns)   StdDev (ns)                             Name
 --------  ---------------  ---------  -----------  -----------  ---------  ---------  -----------  ----------------------------------------------------------
...
      5.7       66,543,457        410    162,301.1     95,008.0     88,127    586,301     83,005.9  dequantize_block_q4_0(int, const void *, float *)
...
```

So, a **≈6.5x** speedup in the **quantization kernel**.

Unfortunately, I don't see any improvement in perplexity calculation times, presumably because even if we finish quantization early, we still have to wait until the second host-to-device transfer finishes. Here is a screenshot of what's happening before the PR:
![Screenshot 2023-04-29 at 00 24 21](https://user-images.githubusercontent.com/42850/235271113-583ce7b8-7c02-481e-a87f-8fc51b2e0db6.png)

And after the PR:
![Screenshot 2023-04-29 at 00 24 52](https://user-images.githubusercontent.com/42850/235271166-bab6240a-1275-4918-ad4e-c8aa4f79e951.png)

However, the quantization kernel becomes consistently faster across various GPUs. Some additional data points below.

**GeForce RTX 3070**

```
BEFORE
 Time (%)  Total Time (ns)  Instances   Avg (ns)     Med (ns)    Min (ns)    Max (ns)   StdDev (ns)                                                  Name
 --------  ---------------  ---------  -----------  -----------  ---------  ----------  -----------  ----------------------------------------------------------------------------------------------------
...
     41,0      835 572 902        342  2 443 195,0  1 415 739,0  1 404 540   8 822 843  1 229 082,0  dequantize_block_q4_0(const void *, float *)
...

AFTER
 Time (%)  Total Time (ns)  Instances   Avg (ns)     Med (ns)    Min (ns)    Max (ns)   StdDev (ns)                                                  Name
 --------  ---------------  ---------  -----------  -----------  ---------  ----------  -----------  ----------------------------------------------------------------------------------------------------
...
     13,0      226 221 174        420    538 621,0    310 908,0    309 132   1 939 892    269 331,0  dequantize_block_q4_0(int, const void *, float *)
...
```

**Tesla M40**

```
BEFORE
 Time (%)  Total Time (ns)  Instances    Avg (ns)     Med (ns)    Min (ns)    Max (ns)   StdDev (ns)                                                  Name
 --------  ---------------  ---------  ------------  -----------  ---------  ----------  -----------  ----------------------------------------------------------------------------------------------------
     50.4    3,157,715,863        303  10,421,504.5  5,971,304.0  5,952,808  37,317,443  5,268,190.7  dequantize_block_q4_0(const void *, float *)
...

AFTER
 Time (%)  Total Time (ns)  Instances   Avg (ns)     Med (ns)    Min (ns)    Max (ns)   StdDev (ns)                                                  Name
 --------  ---------------  ---------  -----------  -----------  ---------  ----------  -----------  ----------------------------------------------------------------------------------------------------
...
      9.1      332,668,931        322  1,033,133.3    599,592.5    566,342   3,600,173    517,167.3  dequantize_block_q4_0(int, const void *, float *)
...
```