@SlyEcho 2023-04-29
That could explain something.

I took a look at the CL kernels and saw they didn't have any loops in them. Thinking this could be a nice way to improve performace, I changed them to work like the CUDA kernels do.

It was worse, a lot worse. I'm sure they could also be optimized, but we need to account for the "warps" and the "wavefronts". I'm all new to this stuff anyway, but it's been a good learning experience.

On a sidenote, there could also be a way to unify the CL and CUDA/ROCm kernels using some "clever" techniques like how [CLBlast also supports CUDA](https://github.com/CNugteren/CLBlast/blob/master/src/kernels/opencl_to_cuda.h)