@slaren 2023-04-29
Very interesting and very informative as always @dfyz! Do you think there would be any advantage to using `cudaOccupancyMaxPotentialBlockSize` instead of using a fixed number of threads?

My only concern with this is that currently adding new quantization types to cuBLAS is very easy, it's just an iteration of the reference C implementation. This looks a lot more complicated, and I am not sure if it is worth it if the bottleneck is elsewhere anyway. Could we get 90% of the way there just by improving occupancy? I know that the occupancy currently is very low, I had tested it with `cudaOccupancyMaxActiveBlocksPerMultiprocessor` during the initial implementation, but since it didn't affect the overall performance I preferred to keep it simple.

In any case, if this becomes the bottleneck in the future, I think we should definitely do this.