@dfyz 2023-04-30
@slaren 

Yeah, I totally agree with your overall message -- adding new quantization methods (which, judging by the issues/discussions, will keep appearing) by porting reference implementation should be easy, and this is indeed way more complicated than that.

Let's keep this PR around if anyone becomes interested in fast CUDA quantization. It's a draft anyway, so I hope it shouldn't clutter anything up.

> Do you think there would be any advantage to using `cudaOccupancyMaxPotentialBlockSize` instead of using a fixed number of threads?

For kernels of this kind, I don't think it makes much difference -- e.g., it says I should launch 768 threads instead of 1024 for `dequantize_block_q4_0`, which doesn't result in any noticeable change in runtime. But yes, using the result of `cudaOccupancyMaxPotentialBlockSize` instead of hardcoded constants makes the code cleaner. I should do this if this PR becomes relevant.

> Could we get 90% of the way there just by improving occupancy?

I played around with this, and it seems to be one of these cases where "higher occupancy does not always equate to higher performance" (quoting [CUDA Best Practices](https://docs.nvidia.com/cuda/cuda-c-best-practices-guide/index.html#occupancy)).

The original kernel runs in **1.66 milliseconds and 2,494,911 cycles**, and has occupancy of ≈30% (because we only use 1 warp, it is limited by the number of block we can fit on an SM):
![image](https://user-images.githubusercontent.com/42850/235380079-57699234-5483-4f20-882a-bd6f03f7516c.png)

The easiest way to improve occupancy is to just increase the number of threads.

<details>
<summary>
  A hacky patch to do this
</summary>

```diff
diff --git a/ggml-cuda.cu b/ggml-cuda.cu
index 5a2701c..52eca3a 100644
--- a/ggml-cuda.cu
+++ b/ggml-cuda.cu
@@ -4,6 +4,8 @@
 #include <atomic>
 #include "ggml-cuda.h"

+#define THREAD_COUNT 768
+
 typedef uint16_t ggml_fp16_t;
 static_assert(sizeof(__half) == sizeof(ggml_fp16_t), "wrong fp16 size");

@@ -53,10 +55,12 @@ typedef struct {
 } block_q8_0;
 static_assert(sizeof(block_q8_0) == sizeof(float) + QK8_0, "wrong q8_0 block size/padding");

-static __global__ void dequantize_block_q4_0(const void * vx, float * y) {
+static __global__ void dequantize_block_q4_0(int nb, const void * vx, float * y) {
     const block_q4_0 * x = (const block_q4_0 *) vx;

-    const int i = blockIdx.x;
+    const int i = blockIdx.x * THREAD_COUNT + threadIdx.x;
+
+    if (i >= nb) return;

     const float d = x[i].d;

@@ -199,7 +203,8 @@ static __global__ void dequantize_block_q8_0(const void * vx, float * y) {

 void dequantize_row_q4_0_cuda(const void * vx, float * y, int k, cudaStream_t stream) {
     const int nb = k / QK4_0;
-    dequantize_block_q4_0<<<nb, 1, 0, stream>>>(vx, y);
+    const int n_thread_blocks = nb / THREAD_COUNT + (nb % THREAD_COUNT != 0);
+    dequantize_block_q4_0<<<n_thread_blocks, THREAD_COUNT, 0, stream>>>(nb, vx, y);
 }

 void dequantize_row_q4_1_cuda(const void * vx, float * y, int k, cudaStream_t stream) {
```
</details>

We get ≈90% occupancy after this:
![image](https://user-images.githubusercontent.com/42850/235380294-7dae5b8d-99db-4077-a6f8-128d664d9054.png)

However, the results get worse: **2.08 milliseconds and 3,114,724 cycles**.

In contrast, the kernel from this PR (with 768 threads) only achieves ≈70% occupancy:
![image](https://user-images.githubusercontent.com/42850/235380412-1b8506d0-4fb8-4c3d-858a-ad764db96b81.png)

Even though the occupancy is lower, the speed is much, much better: **304.19 microseconds and 449,557 cycles**.

My simplified interpretation of these results: since the kernel is memory-bound, speeding up is possible only if we a) decrease the overall number of memory transactions (by coalescing loads), or b) issue more loads when there's available memory bandwidth (by increasing occupancy). This PR targets a), and it seems like in this case this is the thing that matters.

If you also want to test this, use something like this:
```
sudo /usr/local/NVIDIA-Nsight-Compute/ncu -o report --set full -c 1 ./perplexity -m ~/ggml-model-q4_0.bin -f ~/wiki.test.raw
```

This will save a report about the very first kernel (which happens to be a dequantize kernel) to `report.ncu-rep`, which can then be opened in NSight Compute GUI. By increasing the value after `-c`, you can get more kernel launches in the report (but I think the first one is enough for performance analysis).