@dfyz 2023-04-30
@SlyEcho 

> On a sidenote, there could also be a way to unify the CL and CUDA/ROCm kernels using some "clever" techniques

Oh wow, I managed to totally miss the fact that we also have OpenCL dequantize kernels. I'm also new to OpenCL, but based on [this](https://github.com/ProjectPhysX/FluidX3D#solving-the-compatibility-problem), it supports pretty much every GPU imaginable for computations. Maybe we can just drop the custom CUDA/ROCm kernels and replace the with OpenCL ones?