@JohannesGaessler 2023-05-06
I have a GTX 1070. It seems that on this GPU increasing occupancy _does_ lead to better performance. I did a low-effort PR https://github.com/ggerganov/llama.cpp/pull/1341 that increased the number of threads per block to 256. This gave me ~14% faster prompt processing with 33b and a prompt with 399 tokens. When I re-ran the tests with the kernel in this PR prompt processing was ~57% faster compared to master.