@dfyz 2023-05-06
> When I re-ran the tests with the kernel in this PR prompt processing was ~57% faster compared to master.

Whoa, this is nice to hear. Unfortunately I don't have a GTX 1070, but if there is interest, I can try speeding up other kernels in similar manner. I don't know if it's worth it, since it does make the code uglier and harder to iterate on.

(#1341 should definitely be merged independently of this PR, since #1341 is much simpler and cleaner than my approach)