@slaren 2023-05-06
I am in favour of merging anything that improves overall performance, if there are significant gains to be had doing this I think it is great and should be done. In some mat muls dequantization is also a bottleneck in my 3080 too (for example see the timeline in #1269), though overall the different seems small. 

There may be some changes to the quantization formats happening soon (#1305, #1073) though, so it may be better to wait for those.